requirements.txt
watchdog==6.0.0
PyPDF2==3.0.1
python-dotenv==1.2.2
winotify==1.1.0
